src/test/java/ch/supsi/dti/isin/benchmark/adapter/consistenthash/jump/JumpFactoryTests.java
package ch.supsi.dti.isin.benchmark.adapter.consistenthash.jump;

import static org.junit.jupiter.api.Assertions.assertNotNull;

import java.util.function.Supplier;

import org.junit.jupiter.api.DisplayNameGeneration;
import org.junit.jupiter.api.DisplayNameGenerator;
import org.junit.jupiter.api.Test;

import ch.supsi.dti.isin.benchmark.adapter.ConsistentHashFactoryContract;
import ch.supsi.dti.isin.benchmark.config.AlgorithmConfig;
import ch.supsi.dti.isin.consistenthash.jump.JumpEngine;

/**
 * Suite to test the {@link JumpFactory} class.
 * 
 * @author Massimo Coluzzi
 */
@DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
public class JumpFactoryTests implements ConsistentHashFactoryContract<JumpFactory>
{


    /* ******************* */
    /*  INTERFACE METHODS  */
    /* ******************* */


    /**
     * {@inheritDoc}
     */
    @Override
    public JumpFactory sampleValue( AlgorithmConfig config )
    {

        return new JumpFactory( config );

    }


    /* ************** */
    /*  TEST METHODS  */
    /* ************** */


    @Test
    public void default_configuration_should_return_default_engine()
    {

        final JumpFactory factory = sampleValue( CONFIG );
        
        final Supplier<JumpEngine> supplier = factory.createEngineInitializer( FUNCTION, NODES );
        assertNotNull( supplier );

        final JumpEngine engine = supplier.get();
        assertNotNull( engine );

    }

}
